JSAPI: Refactor IJsonWebSocketServer -> IJSWebSocketServer
