+++ Snowflake.Tests/Service/WebSocketServer/JSWebSocketServerTests.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Xunit;
+
+namespace Snowflake.Service.WebSocketServer.Tests
+{
+    public class JSWebSocketServerTests
+    {
+        //todo write integration tests for this
+    }
+}
