+++ Snowflake.Tests/Snowflake.Tests.csproj
@@ -90,6 +90,7 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Scraper\GameImagesResultTests.cs" />
     <Compile Include="Service\HttpServer\BaseHttpServerTests.cs" />
+    <Compile Include="Service\WebSocketServer\JSWebSocketServerTests.cs" />
     <Compile Include="TestUtilities.cs" />
     <Compile Include="Utility\BiDictionaryTests.cs" />
     <Compile Include="Utility\HashTests.cs" />
